Added wow calculation on dashboard
