+++ src/app/(dashboard)/overview/page.tsx
@@ -6,12 +6,12 @@ import { ScrollArea } from "@/components/ui/scroll-area";
 import { useSession } from "next-auth/react";
 import { ContributorDto } from "@/app/api/contributors/fetchUserContributors";
 import { TopContributorsDataTable } from "./top-contributors-data-table";
+import { DashboardInsights } from "./dashboardInsights";
 import { topContributorsColumns } from "./top-contributors-columns";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import { useEffect, useState } from "react";
 import axios from "axios";
 import { Icons } from "@/components/icons";
-import { RegisteredGitRepo } from "@/app/api/github/repo/all/fetchAllUserRepos";
 import { driver } from "driver.js";
 import "driver.js/dist/driver.css";
 import { toast } from "sonner";
@@ -22,8 +22,6 @@ export default function OverviewPage() {
   const [isLoading, setIsLoading] = useState(true);
   const [contributors, setContributors] = useState<ContributorDto[]>([]);
   const [topContributors, setTopContributors] = useState<ContributorDto[]>([]);
-  const [credSum, setCredSum] = useState<number>(0);
-  const [userRepos, setUserRepos] = useState<RegisteredGitRepo[]>([]);
   const [hasSeenProductTour, setHasSeenProductTour] = useState(true);
   const [overview, setOverview] = useState<OverviewDto>();
 
@@ -31,8 +29,6 @@ export default function OverviewPage() {
     if (session?.githubLogin) {
       setIsLoading(false);
       handleFetchContributors();
-      handleFetchCredSum();
-      handleFetchUserRepos();
       handleFetchProductTour();
       handleFetchOverview();
     }
@@ -45,24 +41,10 @@ export default function OverviewPage() {
     }
   };
 
-  const handleFetchCredSum = async () => {
-    const { data } = await axios.get("/api/cred/sum");
-    if (data.success) {
-      setCredSum(data.credSum);
-    }
-  };
-
   const handleFetchOverview = async () => {
     const { data } = await axios.get("/api/overview");
     if (data.success) {
       setOverview(data.overview);
-    }
-  };
-
-  const handleFetchUserRepos = async () => {
-    const { data } = await axios.get("/api/github/repo/all");
-    if (data.success) {
-      setUserRepos(data.registeredRepos);
     }
   };
 
@@ -190,59 +172,8 @@ export default function OverviewPage() {
               </Alert>
             </div>
 
-            <div className="grid gap-4 md:grid-cols-2 lg:grid-cols-3">
-              <Card>
-                <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-                  <CardTitle className="text-sm font-medium">
-                    Total CRED earned
-                  </CardTitle>
-                  <Icons.pizza></Icons.pizza>
-                </CardHeader>
-                <CardContent>
-                  <div className="text-2xl font-bold">
-                    {overview?.totalScore.toFixed(2)}
-                  </div>
-                  <p className="pt-1 text-xs text-muted-foreground">
-                    All CRED earned on all your teams
-                  </p>
-                </CardContent>
-              </Card>
+            <DashboardInsights overview={overview}></DashboardInsights>
 
-              <Card>
-                <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-                  <CardTitle className="text-sm font-medium">
-                    This week CRED earned
-                  </CardTitle>
-                  <Icons.gitBranch></Icons.gitBranch>
-                </CardHeader>
-                <CardContent>
-                  <div className="text-3xl font-bold">
-                    {overview?.weekScore.toFixed(2)}
-                  </div>
-                  <p className="pt-1 text-xs text-muted-foreground">
-                    +{overview?.weekGrowth.toFixed(2)}% from last week
-                  </p>
-                </CardContent>
-              </Card>
-
-              <Card>
-                <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-                  <CardTitle className="text-sm font-medium">
-                    Total contributors
-                  </CardTitle>
-                  <Icons.users></Icons.users>
-                </CardHeader>
-                <CardContent>
-                  <div className="text-2xl font-bold">
-                    {contributors.length}
-                  </div>
-                  <p className="pt-1 text-xs text-muted-foreground">
-                    {/* +15% from last month */}
-                    All contributors on all your teams
-                  </p>
-                </CardContent>
-              </Card>
-            </div>
             <div className="grid gap-4 grid-cols-1 md:grid-cols-2 lg:grid-cols-7">
               <Card className="col-span-7">
                 <CardHeader>
